# front/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Booky API",
        "description": "API for Booky - Book management and social reading platform",
        "contact": {
            "name": "Booky Team",
            "url": "https://booky.com",
            "email": "support@booky.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://choosealicense.com/licenses/mit/"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Development Server"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
        "/users": {
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "Update user profile",
                "description": "Updates user profile information and optionally uploads a profile image",
                "operationId": "updateUserWithImage",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user"
                                ],
                                "type": "object",
                                "properties": {
                                    "user": {
                                        "$ref": "#/components/schemas/UserUpdateDto"
                                    },
                                    "image": {
                                        "type": "string",
                                        "description": "Profile image file (optional)",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/users/unfollow": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Unfollow user",
                "description": "Removes a follow relationship with the specified user",
                "operationId": "unfollowUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FollowUserDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "Unfollow successful"
                    },
                    "400": {
                        "description": "Unfollow failed (user was not being followed)"
                    }
                }
            }
        },
        "/users/search-by-books": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Search users by books for exchange",
                "description": "Find users who have specific books available for exchange, ordered by distance if requesting user has address",
                "operationId": "searchUsersByBooks",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SearchUsersByBooksDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "401": {
                        "description": "Not authenticated"
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "200": {
                        "description": "Users found successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/users/follow": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Follow user",
                "description": "Creates a follow relationship with the specified user",
                "operationId": "followUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FollowUserDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Follow successful"
                    },
                    "400": {
                        "description": "Follow failed (user already followed or invalid ID)"
                    }
                }
            }
        },
        "/sign-up": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Register new user",
                "description": "Creates a new user account with the provided information",
                "operationId": "signUp",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignUpDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Invalid input data or user already exists"
                    },
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sign-in": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "User login",
                "description": "Authenticates a user with email and password",
                "operationId": "signIn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignInDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSignInResponseDto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/users/{id}": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Get user by ID",
                "description": "Retrieves a user's profile information by their unique ID",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "Delete user",
                "description": "Permanently deletes a user account",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/users/{id}/following": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Get users followed by user",
                "description": "Retrieves the list of users that the specified user is following",
                "operationId": "getFollowing",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Following list retrieved successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/users/{id}/followers": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Get user followers",
                "description": "Retrieves the list of users following the specified user",
                "operationId": "getFollowers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Followers retrieved successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AddressDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "state": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    },
                    "longitude": {
                        "type": "number",
                        "format": "double"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
            "UserUpdateDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/AddressDto"
                    }
                },
                "description": "User update data"
            },
            "UserDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/AddressDto"
                    },
                    "date_created": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "FollowUserDto": {
                "required": [
                    "target_user_id"
                ],
                "type": "object",
                "properties": {
                    "target_user_id": {
                        "type": "string",
                        "description": "ID of the user to follow/unfollow",
                        "example": "user-123"
                    }
                },
                "description": "Follow request data"
            },
            "SearchUsersByBooksDto": {
                "required": [
                    "book_ids"
                ],
                "type": "object",
                "properties": {
                    "book_ids": {
                        "maxItems": 10,
                        "minItems": 0,
                        "type": "array",
                        "description": "List of book IDs to search for",
                        "example": [
                            "book-001",
                            "book-002"
                        ],
                        "items": {
                            "type": "string",
                            "description": "List of book IDs to search for",
                            "example": "[\"book-001\",\"book-002\"]"
                        }
                    }
                },
                "description": "Search criteria with book list"
            },
            "UserSignUpDto": {
                "required": [
                    "email",
                    "lastname",
                    "name",
                    "password",
                    "username"
                ],
                "type": "object",
                "properties": {
                    "username": {
                        "maxLength": 30,
                        "minLength": 3,
                        "type": "string",
                        "description": "Unique username",
                        "example": "johndoe123"
                    },
                    "name": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string",
                        "description": "User's first name",
                        "example": "John"
                    },
                    "lastname": {
                        "maxLength": 50,
                        "minLength": 2,
                        "type": "string",
                        "description": "User's last name",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "description": "User's email address",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "maxLength": 100,
                        "minLength": 8,
                        "type": "string",
                        "description": "User's password",
                        "example": "SecurePassword123!"
                    }
                },
                "description": "User registration data"
            },
            "UserSignInDto": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "User's email address",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "description": "User's password",
                        "example": "SecurePassword123!"
                    }
                },
                "description": "User login credentials"
            },
            "UserSignInResponseDto": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "JWT authentication token",
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserDto"
                    }
                },
                "description": "User sign-in response with user data and JWT token"
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}